requirements.txt
absl-py==0.12.0
addict==2.4.0
adict==0.1.7
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.1.0
ci-info==0.2.0
click==8.1.3
cloudpickle==2.0.0
cryptography==35.0.0
cycler==0.10.0
Cython==0.29.23
decorator==5.0.9
dominate==2.6.0
etelemetry==0+unknown
filelock==3.7.1
fonttools==4.28.2
frozenlist==1.3.1
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.30.0
google-auth-oauthlib==0.4.4
GPUtil==1.4.0
grpcio==1.37.1
h5py==3.3.0
html5lib==1.1
idna==2.7
imageio==2.4.1
importlib-metadata==4.11.4
ipython==7.23.1
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.18.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.2.1
loguru==0.4.1
looseversion==1.0.1
lxml==4.8.0
Markdown==3.3.4
matplotlib==3.5.0
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
networkx==2.6.3
neurite==0.2
nibabel==3.2.2
nipype==1.8.3
numpy==1.21.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.2.52
packaging==21.0
pandas==1.3.4
parso==0.8.2
pbr==5.8.0
pexpect==4.8.0
pgi==0.0.11.2
pickleshare==0.7.5
Pillow==8.2.0
pip==21.2.4
progressbar2==3.38.0
prompt-toolkit==3.0.18
protobuf==3.17.0
prov==2.0.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydicom==2.3.0
pydot==1.2.4
Pygments==2.9.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
pystrum==0.2
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.0.3
rarfile==4.0
rdflib==6.1.1
requests==2.19.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.14.5
scikit-learn==0.19.2
scipy==1.6.3
setuptools==58.0.4
setuptools-scm==6.3.2
simplejson==3.17.6
six==1.16.0
smmap==4.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
testresources==2.0.1
tf-slim==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
tomli==1.2.2
torch==1.9.0
torchaudio==0.8.0a0+e4e171a
torchfile==0.1.0
torchstat==0.0.7
torchvision==0.10.0
tornado==6.1
tqdm==4.25.0
traitlets==5.0.5
traits==6.2.0
tta==0.2.1
ttach==0.0.3
typing-extensions==3.10.0.0
urllib3==1.23
visdom==0.1.8.9
vtk==9.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.0.0
wheel==0.37.0
wrapt==1.10.11
wslink==1.8.2
xlrd==2.0.1
xvfbwrapper==0.2.9
yapf==0.31.0
yarl==1.8.1
zipp==3.8.0
